ui(refactor): split authoring module into multiple files
