+++ scripts/superdesk-authoring/authoring/authoring.spec.js
