+++ scripts/superdesk-authoring/authoring/controllers/index.js
@@ -1,0 +1,2 @@
+export { AuthoringController } from './AuthoringController';
+export { ChangeImageController } from './ChangeImageController';
