+++ scripts/superdesk-authoring/authoring/directives/Html5vfix.js
@@ -1,0 +1,11 @@
+/**
+ * Html5 Source tag doesn't support {{ angular }}
+ */
+export function Html5vfix() {
+    return {
+        restrict: 'A',
+        link: function(scope, element, attr) {
+            attr.$set('src', attr.vsrc);
+        }
+    };
+}
