+++ scripts/superdesk-authoring/authoring/directives/RemoveTagsDirective.js
@@ -1,0 +1,19 @@
+import * as helpers from 'superdesk-authoring/authoring/helpers';
+
+RemoveTagsDirective.$inject = [];
+export function RemoveTagsDirective() {
+    var htmlRegex = /(<([^>]+)>)/ig;
+    return {
+        require: 'ngModel',
+        scope: {
+            model: '=ngModel'
+        },
+        link: function(scope, elem, attr, ngModel) {
+            scope.$watch('model', function() {
+                if (scope.model) {
+                    scope.model = helpers.stripHtmlRaw(scope.model).replace(htmlRegex, '');
+                }
+            });
+        }
+    };
+}
