+++ scripts/superdesk-authoring/authoring/filters/index.js
@@ -1,0 +1,13 @@
+export function EmbeddedFilter() {
+    return function(input) {
+        var output = {};
+        for (var i in input) {
+            if (input.hasOwnProperty(i)) {
+                if (!/^embedded/.test(i)) {
+                    output[i] = input[i];
+                }
+            }
+        }
+        return output;
+    };
+}
