+++ scripts/superdesk-authoring/authoring/services/RenditionsService.js
@@ -1,56 +1,5 @@
-angular.module('superdesk.authoring')
-    .service('renditions', RenditionsService)
-    .factory('history', HistoryFactory);
-
-/**
- * Watches an expression to keep history of its states
- * and binds ctrl-z and ctrl-y to undo/redo its states
- */
-HistoryFactory.$inject = ['History', '$window', '$timeout'];
-function HistoryFactory(History, $window, $timeout) {
-    var KeyOperations = {
-        ['Z'.charCodeAt(0)]: History.undo,
-        ['Y'.charCodeAt(0)]: History.redo
-    };
-    return {
-        watch: function(expression, scope) {
-            $timeout(function() {
-                History.watch(expression, scope);
-            }, 0, false);
-            var onHistoryKey = function(event, cb) {
-                var modifier = event.ctrlKey || event.metaKey;
-                if (modifier && KeyOperations[event.keyCode]) {
-                    cb();
-                }
-            };
-            var onHistoryKeydown = function(event) {
-                onHistoryKey(event, function() {
-                    event.preventDefault();
-                    // action is on keydown becuase command key (event.metakey) on OSX is not detected on keyup events
-                    // for some reason.
-                    scope.$apply(function() {
-                        KeyOperations[event.keyCode].bind(History)(expression, scope);
-                    });
-                });
-            };
-            var onHistoryKeyup = function(event) {
-                onHistoryKey(event, function() {
-                    event.preventDefault();
-                });
-            };
-            angular.element($window).on('keydown', onHistoryKeydown);
-            angular.element($window).on('keyup', onHistoryKeyup);
-            scope.$on('$destroy', function() {
-                angular.element($window).unbind('keydown', onHistoryKeydown);
-                angular.element($window).unbind('keyup', onHistoryKeyup);
-                History.forget(scope, expression);
-            });
-        }
-    };
-}
-
 RenditionsService.$inject = ['metadata', '$q', 'api', 'superdesk', 'lodash'];
-function RenditionsService(metadata, $q, api, superdesk, _) {
+export function RenditionsService(metadata, $q, api, superdesk, _) {
     var self = this;
     this.ingest = function(item) {
         var performRenditions = $q.when(item);
