+++ scripts/superdesk-authoring/authoring/services/index.js
@@ -1,0 +1,8 @@
+export { AutosaveService } from './AutosaveService';
+export { AuthoringService } from './AuthoringService';
+export { ConfirmDirtyService } from './ConfirmDirtyService';
+export { LockService } from './LockService';
+export { AuthoringThemesService } from './AuthoringThemesService';
+export { AuthoringWorkspaceService } from './AuthoringWorkspaceService';
+export { RenditionsService } from './RenditionsService';
+export { HistoryFactory } from './HistoryFactory';
